Code completion in Micronaut config yaml files available via LSP.
