+++ enterprise/micronaut/src/org/netbeans/modules/micronaut/MicronautConfigCompletionTask.java
@@ -1,0 +1,205 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.netbeans.modules.micronaut;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import javax.swing.text.BadLocationException;
+import javax.swing.text.Document;
+import org.netbeans.api.editor.document.LineDocument;
+import org.netbeans.api.editor.document.LineDocumentUtils;
+import org.netbeans.modules.csl.api.StructureItem;
+import org.netbeans.modules.csl.api.StructureScanner;
+import org.netbeans.modules.csl.core.Language;
+import org.netbeans.modules.csl.core.LanguageRegistry;
+import org.netbeans.modules.csl.spi.ParserResult;
+import org.netbeans.modules.editor.indent.api.IndentUtils;
+import org.netbeans.modules.parsing.api.ParserManager;
+import org.netbeans.modules.parsing.api.ResultIterator;
+import org.netbeans.modules.parsing.api.Source;
+import org.netbeans.modules.parsing.api.UserTask;
+import org.netbeans.modules.parsing.spi.Parser;
+import org.openide.util.Exceptions;
+import org.springframework.boot.configurationmetadata.ConfigurationMetadataProperty;
+
+/**
+ *
+ * @author Dusan Balek
+ */
+public class MicronautConfigCompletionTask {
+
+    public static interface ItemFactory<T> {
+        T createTopLevelPropertyItem(String propName, int offset, int baseIndent, int indentLevelSize);
+        T createPropertyItem(ConfigurationMetadataProperty property, int offset, int baseIndent, int indentLevelSize, int idx);
+    }
+
+    public <T> List<T> query(Document doc, int caretOffset, MicronautConfigProperties configProperties, ItemFactory<T> factory) {
+        List<T> items = new ArrayList<>();
+        LineDocument lineDocument = LineDocumentUtils.as(doc, LineDocument.class);
+        if (lineDocument != null) {
+            int lineStart = LineDocumentUtils.getLineStart(lineDocument, caretOffset);
+            try {
+                String text = lineDocument.getText(lineStart, caretOffset - lineStart);
+                if (!text.contains("#")) {
+                    int idx = text.indexOf(':');
+                    if (idx < 0) {
+                        final int lineIndent = Math.min(IndentUtils.lineIndent(lineDocument, lineStart), caretOffset - lineStart);
+                        final int wordStart = LineDocumentUtils.getPreviousWhitespace(lineDocument, caretOffset) + 1;
+                        final String prefix = wordStart < caretOffset ? text.substring(wordStart - lineStart) : null;
+                        final int anchorOffset = wordStart < caretOffset ? wordStart : caretOffset;
+                        ParserManager.parse(Collections.singleton(Source.create(lineDocument)), new UserTask() {
+                            public @Override void run(ResultIterator resultIterator) throws Exception {
+                                Parser.Result r = resultIterator.getParserResult();
+                                if (r instanceof ParserResult) {
+                                    Language language = LanguageRegistry.getInstance().getLanguageByMimeType(resultIterator.getSnapshot().getMimeType());
+                                    if (language != null) {
+                                        StructureScanner scanner = language.getStructure();
+                                        if (scanner != null) {
+                                            List<? extends StructureItem> structures = scanner.scan((ParserResult) r);
+                                            int indentLevelSize = getIndentLevelSize(lineDocument, structures);
+                                            List<? extends StructureItem> context = getContext(structures, wordStart);
+                                            String filter = "";
+                                            StructureItem currentItem = null;
+                                            for (StructureItem item : context) {
+                                                int itemLineStart = LineDocumentUtils.getLineStart(lineDocument, (int) item.getPosition());
+                                                int itemLineIndent = IndentUtils.lineIndent(lineDocument, itemLineStart);
+                                                if (itemLineIndent < lineIndent) {
+                                                    filter += item.getName() + '.';
+                                                    currentItem = item;
+                                                }
+                                            }
+                                            int filterLength = filter.length();
+                                            if (prefix != null) {
+                                                filter += prefix;
+                                            }
+                                            int currentItemLineIndent = currentItem != null ? IndentUtils.lineIndent(lineDocument, LineDocumentUtils.getLineStart(lineDocument, (int) currentItem.getPosition())) : -indentLevelSize;
+                                            Map<String, ConfigurationMetadataProperty> properties = configProperties.getProperties();
+                                            Set<String> topLevels = new HashSet<>();
+                                            for (Map.Entry<String, ConfigurationMetadataProperty> entry : properties.entrySet()) {
+                                                String propName = entry.getKey();
+                                                int idx = match(propName, filter);
+                                                if (idx >= 0) {
+                                                    int dotIdx = propName.indexOf('.', idx);
+                                                    String simpleName = dotIdx < 0 ? propName.substring(idx) : propName.substring(idx, dotIdx);
+                                                    if (filter(currentItem != null ? currentItem.getNestedItems() : structures, simpleName)) {
+                                                        items.add(factory.createPropertyItem(entry.getValue(), anchorOffset, currentItemLineIndent + indentLevelSize, indentLevelSize, idx));
+                                                        if (dotIdx > 0) {
+                                                            topLevels.add(simpleName);
+                                                        }
+                                                    }
+                                                }
+                                            }
+                                            for (String topLevel : topLevels) {
+                                                items.add(factory.createTopLevelPropertyItem(topLevel, anchorOffset, currentItemLineIndent + indentLevelSize, indentLevelSize));
+                                            }
+                                        }
+                                    }
+                                }
+                            }
+                        });
+                    }
+                }
+            } catch (Exception ex) {
+                Exceptions.printStackTrace(ex);
+            }
+        }
+        return items;
+    }
+
+    private int match(String propName, String filter) {
+        int len = 0;
+        String[] parts = propName.split("\\*");
+        for (int i = 0; i < parts.length; i++) {
+            String part = parts[i];
+            if (filter.length() <= part.length()) {
+                if (part.startsWith(filter)) {
+                    int idx = filter.lastIndexOf('.');
+                    return idx < 0 ? len : len + idx + 1;
+                } else {
+                    return -1;
+                }
+            }
+            if (i < parts.length - 1) {
+                if (!part.equals(filter.substring(0, part.length()))) {
+                    return -1;
+                }
+                len += part.length();
+                int idx = filter.indexOf('.', part.length());
+                if (idx < 0) {
+                    return -1;
+                }
+                len += 1;
+                filter = filter.substring(idx);
+            } else {
+                return filter.startsWith(part) ? len : -1;
+            }
+        }
+        return -1;
+    }
+
+    private boolean filter(List<? extends StructureItem> structures, String name) {
+        for (StructureItem item : structures) {
+            if (name.equals(item.getName())) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    private List<StructureItem> getContext(List<? extends StructureItem> structure, int offset) {
+        List<StructureItem> items = new ArrayList<>();
+        while (structure != null && !structure.isEmpty()) {
+            StructureItem currentItem = null;
+            for (StructureItem item : structure) {
+                if (item.getPosition() < offset) {
+                    currentItem = item;
+                }
+            }
+            if (currentItem != null) {
+                items.add(currentItem);
+                structure = currentItem.getNestedItems();
+            } else {
+                structure = null;
+            }
+        }
+        return items;
+    }
+
+    private int getIndentLevelSize(LineDocument lineDocument, List<? extends StructureItem> structures) {
+        int indentLevel = IndentUtils.indentLevelSize(lineDocument);
+        try {
+            for (StructureItem structure : structures) {
+                int baseStart = LineDocumentUtils.getLineStart(lineDocument, (int) structure.getPosition());
+                int baseIndent = IndentUtils.lineIndent(lineDocument, baseStart);
+                for (StructureItem nestedItem : structure.getNestedItems()) {
+                    int lineStart = LineDocumentUtils.getLineStart(lineDocument, (int) nestedItem.getPosition());
+                    int lineIndent = IndentUtils.lineIndent(lineDocument, lineStart) - baseIndent;
+                    if (lineIndent > 0 && lineIndent < indentLevel) {
+                        indentLevel = lineIndent;
+                    }
+                }
+            }
+        } catch (BadLocationException ble) {}
+        return indentLevel;
+    }
+}
