+++ ide/editor.completion/apichanges.xml
@@ -84,6 +84,21 @@ is the proper place.
     <!-- ACTUAL CHANGES BEGIN HERE: -->
 
     <changes>
+        <change id="CompletionProvider.getCompletionItems">
+            <api name="completion"/>
+            <summary>Addition of CompletionProvider.getCompletionItems()</summary>
+            <version major="1" minor="57"/>
+            <date day="16" month="3" year="2021"/>
+            <author login="dbalek"/>
+            <compatibility addition="yes"/>
+            <description>
+            <p>
+                <code>CompletionProvider.getCompletionItems()</code> method was added to allow for
+                completion items computation and sending them to outside NetBeans clients for
+                display (e.g. via Language Server Protocol).
+            </p>
+            </description>
+        </change>
         <change id="CompositeCompletionItem">
             <api name="completion"/>
             <summary>Addition of CompositeCompletionItem</summary>
