+++ ide/editor.completion/nbproject/project.properties
@@ -19,4 +19,4 @@ javac.compilerargs=-Xlint:unchecked
 javac.source=1.8
 javadoc.arch=${basedir}/arch.xml
 javadoc.apichanges=${basedir}/apichanges.xml
-spec.version.base=1.56.0
+spec.version.base=1.57.0
